requirements.txt
pandas
Pillow
matplotlib
numpy
openslide-python
torch
opencv-python
scikit-image